requirements.txt
requests
Pillow
xxhash
redis